requirements.txt
# Dev environment
pip
autopep8

starlette
python-fasthtml
fastlite
MonsterUI